Update ActuatorToLogsEmitter.java
The type of sample.value shoud be Integer rather than Float. Otherwise, Metric Registrar can't scrape COUNTER value metrics correctly.

+++ java-spring-actuator-metrics-to-structured-logs/src/main/java/io/pivotal/metric_registrar/examples/spring_actuator_to_logs/ActuatorToLogsEmitter.java
@@ -48,7 +48,7 @@ public class ActuatorToLogsEmitter {
             case GAUGE:
                 return String.format("{ \"type\": \"gauge\", \"value\": %s, %s", sample.value, nameAndTags);
             case COUNTER:
-                return String.format("{ \"type\": \"counter\", \"delta\": %s, %s", sample.value, nameAndTags);
+                return String.format("{ \"type\": \"counter\", \"delta\": %s, %s", (int)(sample.value), nameAndTags);
             default:
                 return null;
         }
